# gwen-generated-api/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DTOGameStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"FINISHED",
					"ABANDONED"
				]
			},
			"DTOGame": {
				"properties": {
					"updated_at": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"winner_id": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/DTOGameStatus"
					},
					"player2_id": {
						"type": "string"
					},
					"player1_id": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"winner_id",
					"status",
					"player2_id",
					"player1_id",
					"_id"
				],
				"type": "object"
			},
			"DTOGamePhase": {
				"type": "string",
				"enum": [
					"WAITING_FOR_PLAYERS",
					"REDRAW",
					"FLIP_COIN",
					"PLAY_CARDS",
					"END"
				]
			},
			"DTOPlayableCard": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DTODeck": {
				"properties": {
					"drawPile": {
						"items": {
							"$ref": "#/components/schemas/DTOPlayableCard"
						},
						"type": "array"
					},
					"hand": {
						"items": {
							"$ref": "#/components/schemas/DTOPlayableCard"
						},
						"type": "array"
					},
					"discardPile": {
						"items": {
							"$ref": "#/components/schemas/DTOPlayableCard"
						},
						"type": "array"
					},
					"leader": {
						"$ref": "#/components/schemas/DTOPlayableCard"
					},
					"factionId": {
						"type": "string"
					}
				},
				"required": [
					"drawPile",
					"hand",
					"discardPile",
					"leader",
					"factionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DTOPlayer": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"gems": {
						"type": "number",
						"format": "double"
					},
					"passed": {
						"type": "boolean"
					},
					"deck": {
						"$ref": "#/components/schemas/DTODeck"
					}
				},
				"required": [
					"userId",
					"gems",
					"passed",
					"deck"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RangeType": {
				"type": "string",
				"enum": [
					"MELEE",
					"RANGED",
					"SIEGE"
				]
			},
			"DTORow": {
				"properties": {
					"score": {
						"type": "number",
						"format": "double"
					},
					"cardsIds": {
						"items": {
							"$ref": "#/components/schemas/DTOPlayableCard"
						},
						"type": "array"
					},
					"range": {
						"$ref": "#/components/schemas/RangeType"
					}
				},
				"required": [
					"score",
					"cardsIds",
					"range"
				],
				"type": "object"
			},
			"DTOPlayerRows": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/DTORow"
						},
						"type": "array"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"rows",
					"score",
					"userId"
				],
				"type": "object"
			},
			"DTOGameResult": {
				"type": "string",
				"enum": [
					"WIN",
					"LOSS",
					"DRAW"
				]
			},
			"DTORoundEndResult": {
				"properties": {
					"player2_gems": {
						"type": "number",
						"format": "double"
					},
					"player1_gems": {
						"type": "number",
						"format": "double"
					},
					"player2_score": {
						"type": "number",
						"format": "double"
					},
					"player1_score": {
						"type": "number",
						"format": "double"
					},
					"player2_result": {
						"$ref": "#/components/schemas/DTOGameResult"
					},
					"player1_result": {
						"$ref": "#/components/schemas/DTOGameResult"
					},
					"round": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"player2_gems",
					"player1_gems",
					"player2_score",
					"player1_score",
					"player2_result",
					"player1_result",
					"round"
				],
				"type": "object",
				"description": "Round end event: Shows overlay for ~3 seconds during game\nContains each player's result from their perspective"
			},
			"DTOGameEndResult": {
				"properties": {
					"player2_elo_change": {
						"type": "number",
						"format": "double"
					},
					"player1_elo_change": {
						"type": "number",
						"format": "double",
						"description": "TODO: Will be populated when Elo system is implemented"
					},
					"winner_id": {
						"type": "string"
					},
					"player2_gems_lost": {
						"type": "number",
						"format": "double"
					},
					"player1_gems_lost": {
						"type": "number",
						"format": "double"
					},
					"player2_result": {
						"$ref": "#/components/schemas/DTOGameResult"
					},
					"player1_result": {
						"$ref": "#/components/schemas/DTOGameResult"
					},
					"player2_id": {
						"type": "string"
					},
					"player1_id": {
						"type": "string"
					}
				},
				"required": [
					"winner_id",
					"player2_gems_lost",
					"player1_gems_lost",
					"player2_result",
					"player1_result",
					"player2_id",
					"player1_id"
				],
				"type": "object",
				"description": "Game end result: Shows persistent page when game is over\nIncludes standings and final results"
			},
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"DTOGameWithMetadata": {
				"properties": {
					"usernames": {
						"$ref": "#/components/schemas/Record_string.string_",
						"description": "Mapping of userId to username for display purposes"
					},
					"game": {
						"properties": {
							"gameEndResult": {
								"allOf": [
									{
										"$ref": "#/components/schemas/DTOGameEndResult"
									}
								],
								"nullable": true,
								"description": "Final game result (null if game not ended)"
							},
							"lastRoundResult": {
								"allOf": [
									{
										"$ref": "#/components/schemas/DTORoundEndResult"
									}
								],
								"nullable": true,
								"description": "Result of last completed round (null if no round completed)"
							},
							"player2Rows": {
								"$ref": "#/components/schemas/DTOPlayerRows"
							},
							"player1Rows": {
								"$ref": "#/components/schemas/DTOPlayerRows"
							},
							"player2": {
								"$ref": "#/components/schemas/DTOPlayer"
							},
							"player1": {
								"$ref": "#/components/schemas/DTOPlayer"
							},
							"currentPlayerTurnUserId": {
								"type": "string",
								"nullable": true
							},
							"currentRound": {
								"type": "number",
								"format": "double"
							},
							"phase": {
								"$ref": "#/components/schemas/DTOGamePhase"
							}
						},
						"required": [
							"player2Rows",
							"player1Rows",
							"player2",
							"player1",
							"currentPlayerTurnUserId",
							"currentRound",
							"phase"
						],
						"type": "object"
					},
					"metadata": {
						"$ref": "#/components/schemas/DTOGame"
					}
				},
				"required": [
					"usernames",
					"game",
					"metadata"
				],
				"type": "object"
			},
			"DTOPlaceCardRequest": {
				"properties": {
					"rowType": {
						"type": "string",
						"enum": [
							"MELEE",
							"RANGED",
							"SIEGE"
						]
					},
					"cardId": {
						"type": "string"
					},
					"playerId": {
						"type": "string"
					}
				},
				"required": [
					"rowType",
					"cardId",
					"playerId"
				],
				"type": "object"
			},
			"DTOPassTurnRequest": {
				"properties": {
					"playerId": {
						"type": "string"
					}
				},
				"required": [
					"playerId"
				],
				"type": "object"
			},
			"DTOSurrenderRequest": {
				"properties": {
					"playerId": {
						"type": "string"
					}
				},
				"required": [
					"playerId"
				],
				"type": "object"
			},
			"DTOFinishGameRequest": {
				"properties": {
					"winnerId": {
						"type": "string"
					}
				},
				"required": [
					"winnerId"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Gwen API",
		"version": "1.0.0",
		"description": "Gwen game server API",
		"contact": {
			"name": "Gwen Team"
		}
	},
	"paths": {
...
		"/games/{gameId}/active": {
			"get": {
				"operationId": "GetGameWithMetadataById",
				"responses": {
					"200": {
						"description": "Active game with metadata",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGameWithMetadata"
								}
							}
						}
					},
					"404": {
						"description": "Active game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/games/{gameId}": {
			"get": {
				"operationId": "GetGame",
				"responses": {
					"200": {
						"description": "Game details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGame"
								}
							}
						}
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/games/{gameId}/end-round": {
			"post": {
				"operationId": "EndRound",
				"responses": {
					"200": {
						"description": "Round ended, results calculated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGameWithMetadata"
								}
							}
						}
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/games/{gameId}/start-round": {
			"post": {
				"operationId": "StartRound",
				"responses": {
					"200": {
						"description": "Round started successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGameWithMetadata"
								}
							}
						}
					},
					"400": {
						"description": "Cannot start round"
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/games/{gameId}/place-card": {
			"post": {
				"operationId": "PlaceCard",
				"responses": {
					"200": {
						"description": "Card placed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGameWithMetadata"
								}
							}
						}
					},
					"400": {
						"description": "Invalid placement"
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DTOPlaceCardRequest"
							}
						}
					}
				}
			}
		},
		"/games/{gameId}/pass-turn": {
			"post": {
				"operationId": "PassTurn",
				"responses": {
					"200": {
						"description": "Turn passed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGameWithMetadata"
								}
							}
						}
					},
					"400": {
						"description": "Cannot pass turn"
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DTOPassTurnRequest"
							}
						}
					}
				}
			}
		},
		"/games/{gameId}/resign": {
			"post": {
				"operationId": "ResignGame",
				"responses": {
					"200": {
						"description": "Player resigned, opponent wins",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGameWithMetadata"
								}
							}
						}
					},
					"400": {
						"description": "Missing playerId"
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DTOSurrenderRequest"
							}
						}
					}
				}
			}
		},
		"/games/{gameId}/finish": {
			"post": {
				"operationId": "FinishGame",
				"responses": {
					"200": {
						"description": "Game finished successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGame"
								}
							}
						}
					},
					"400": {
						"description": "Missing winnerId"
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DTOFinishGameRequest"
							}
						}
					}
				}
			}
		},
		"/games/{gameId}/abandon": {
			"post": {
				"operationId": "AbandonGame",
				"responses": {
					"200": {
						"description": "Game abandoned successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DTOGame"
								}
							}
						}
					},
					"404": {
						"description": "Game not found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}